Create ContextAPI - BattleTitleRepository, Service 제목 배틀 동안 제목 인덱스를 임시로 따와서 Web Socket 에서 속도를 향상할 수 있도록 하기 위해 Redis 에 오늘의 요청에 해당되는 제목 인덱스를 형성한 List 자료구조를 생성하였음. 그리고 오늘의 요청에 해당되는 제목을 삽입, 수정, 삭제할 때 이 List 자료구조에서 index 의 이동을 반영하기 위해 Cache Listener 를 생성하였음.

+++ ContextAPI/src/main/java/io/kang/service/integrate_service/implement_object/TitleFetchServiceImpl.java
@@ -3,17 +3,24 @@ package io.kang.service.integrate_service.implement_object;
 import io.kang.dto.mysql.RequestDTO;
 import io.kang.dto.mysql.TitleDTO;
 import io.kang.dto.mysql.TitleEmpathyDTO;
+import io.kang.dto.redis.TodayRequestDTO;
 import io.kang.enumeration.Status;
 import io.kang.model.TitleModel;
+import io.kang.repository.redis.BattleTitleRepository;
+import io.kang.service.domain_service.interfaces.BattleTitleService;
 import io.kang.service.domain_service.interfaces.EmpathyService;
 import io.kang.service.domain_service.interfaces.RequestService;
 import io.kang.service.domain_service.interfaces.TitleService;
+import io.kang.service.domain_service.interfaces.TodayRequestService;
 import io.kang.service.integrate_service.interfaces.TitleFetchService;
+import io.kang.service.integrate_service.listener_interfaces.TitleCacheListener;
+import io.kang.vo.BattleFetchRequestVO;
 import io.kang.vo.MainTitleVO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -28,6 +35,36 @@ public class TitleFetchServiceImpl implements TitleFetchService {
 
     @Autowired
     private EmpathyService<TitleEmpathyDTO, TitleDTO> titleEmpathyService;
+
+    @Autowired
+    private TodayRequestService todayRequestService;
+
+    @Autowired
+    private BattleTitleService battleTitleService;
+
+    private TitleCacheListener titleCacheListener = new TitleCacheListener() {
+        @Override
+        public void onCreateInTodayRequest(final Long titleId) {
+            battleTitleService.create(titleId);
+        }
+
+        @Override
+        public void onUpdateInTodayRequest(final Long titleId) {
+            battleTitleService.deleteByTitleId(titleId);
+            battleTitleService.create(titleId);
+        }
+
+        @Override
+        public void onDeleteInTodayRequest(final Long titleId) {
+            battleTitleService.deleteByTitleId(titleId);
+        }
+    };
+
+    private boolean isBeTodayRequest(RequestDTO requestDTO) throws IOException {
+        TodayRequestDTO todayRequestDTO = todayRequestService.findByLast();
+        if(todayRequestDTO == null || requestDTO == null) return false;
+        else return todayRequestDTO.getRequestId() == requestDTO.getId();
+    }
 
     @Override
     public MainTitleVO fetchUserRequestTitle(final Long requestId, final String userId) {
@@ -54,16 +91,20 @@ public class TitleFetchServiceImpl implements TitleFetchService {
 
     @Override
     @Transactional
-    public boolean executeTitleSaving(final TitleModel titleModel) {
+    public boolean executeTitleSaving(final TitleModel titleModel) throws IOException {
         RequestDTO requestDTO = requestService.findById(titleModel.getRequestId());
         if(requestDTO != null){
             TitleDTO titleDTO = titleService.findByUserIdAndRequest(titleModel.getUserId(), requestDTO);
             if(titleDTO != null){
                 TitleDTO updateTitleDTO = TitleModel.builtToDTOIsExisted(titleDTO.getId(), requestDTO, titleModel, LocalDateTime.now());
-                titleService.update(updateTitleDTO);
+                TitleDTO tmpResult = titleService.update(updateTitleDTO);
+                if(this.isBeTodayRequest(tmpResult.getRequest()))
+                    this.titleCacheListener.onUpdateInTodayRequest(tmpResult.getId());
             } else {
                 TitleDTO insertTitleDTO = TitleModel.builtToDTO(requestDTO, titleModel);
-                titleService.create(insertTitleDTO);
+                TitleDTO tmpResult = titleService.create(insertTitleDTO);
+                if(this.isBeTodayRequest(tmpResult.getRequest()))
+                    this.titleCacheListener.onCreateInTodayRequest(tmpResult.getId());
             }
             return true;
         } else return false;
@@ -71,9 +112,13 @@ public class TitleFetchServiceImpl implements TitleFetchService {
 
     @Override
     @Transactional
-    public boolean executeTitleDeleting(final Long titleId) {
+    public boolean executeTitleDeleting(final Long titleId) throws IOException {
+        TitleDTO titleDTO = titleService.findById(titleId);
+        if(titleDTO == null) return false;
         if(titleService.existsById(titleId)){
             titleService.deleteById(titleId);
+            if(this.isBeTodayRequest(titleDTO.getRequest()))
+                this.titleCacheListener.onDeleteInTodayRequest(titleId);
             return true;
         } else return false;
     }
